requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.4
nltk>=3.8.0
spacy>=3.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
datasets>=2.14.0
statsmodels>=0.14.0

